+++ lib/components/themes.dart
@@ -1,6 +1,6 @@
-import 'package:club_app/components/colors.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:hncc_application/components/colors.dart';
 
 class AppThemeData {
   static ThemeData lightThemeData(BuildContext context) => ThemeData(
