+++ lib/views/login.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:adobe_xd/pinned.dart';
-import 'package:hncc_application/constants/theme_dart.dart';
+import 'package:hncc_application/components/colors.dart';
 import 'package:hncc_application/models/up.dart';
 import 'package:hncc_application/views/signup.dart';
 import './events.dart';
@@ -40,7 +40,7 @@ class Login extends StatelessWidget {
                             decoration: BoxDecoration(
                               borderRadius:
                               BorderRadius.all(Radius.elliptical(9999.0, 9999.0)),
-                              color:CustomColors.background,
+                              color:kLtGrey,
                               boxShadow: [
                                 BoxShadow(
                                   color: const Color(0x8097a7c3),
@@ -64,10 +64,10 @@ class Login extends StatelessWidget {
                             decoration: BoxDecoration(
                               borderRadius:
                               BorderRadius.all(Radius.elliptical(9999.0, 9999.0)),
-                              color: CustomColors.background,
+                              color: kLtGrey,
                               boxShadow: [
                                 BoxShadow(
-                                  color: CustomColors.white,
+                                  color: kLtWhite,
                                   offset: Offset(-10, -10),
                                   blurRadius: 15,
                                 ),
@@ -141,7 +141,7 @@ class Login extends StatelessWidget {
                           // style: TextStyle(
                           //   fontFamily: 'Segoe UI',
                           //   fontSize: 17,
-                          //   color:CustomColors.grey,
+                          //   color:kLtGrey,
                           //   fontWeight: FontWeight.w600,
                           // ),
                           // textAlign: TextAlign.center,
@@ -174,7 +174,7 @@ class Login extends StatelessWidget {
                       style: TextStyle(
                         fontFamily: 'Segoe UI',
                         fontSize: 15,
-                        color:CustomColors.darkGrey,
+                        color:kMdGrey,
                         fontWeight: FontWeight.w600,
                       ),
                       textAlign: TextAlign.left,
@@ -225,7 +225,7 @@ class Login extends StatelessWidget {
                       style: TextStyle(
                         fontFamily: 'Corbel',
                         fontSize: 28,
-                        color: CustomColors.black,
+                        color: kLtBlack,
                         fontWeight: FontWeight.w700,
                       ),
                       textAlign: TextAlign.center,
@@ -249,7 +249,7 @@ class Login extends StatelessWidget {
                         style: TextStyle(
                           fontFamily: 'Corbel',
                           fontSize: 22,
-                          color: CustomColors.black,
+                          color: kLtBlack,
                           fontWeight: FontWeight.w300,
                         ),
                         textAlign: TextAlign.center,
