+++ lib/widgets/cardsAndBg.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:hncc_application/dataModals/DemoData.dart';
+import 'package:hncc_application/models/DemoData.dart';
 
 import '../views/teams.dart';
 import 'buttons.dart';
